auth/requirements.txt
fastapi
uvicorn
passlib
python-jose
pyotp
supabase
python-dotenv
